// src/app/interface/atencion.interface.ts
// Generated by https://quicktype.io

export interface ResultAtenciones {
  ok:       boolean;
  msg:      string;
  atencion: Atencion[];
}
export interface ResultAtencion {
  ok:       boolean;
  msg:      string;
  atencion: Atencion;
}
export interface Atencion {
  id:               number;
  id_directorio:    number;
  id_horario:       number;
  id_tipo_atencion: number;
  descripcion:      string;
  Directorio:       Directorio;
  Horario:          Horario;
  TipoAtencion:     Horario;
}

export interface Directorio {
  nombre: string;
  id:     number;
}

export interface Horario {
  descripcion: string;
}